pyproject.toml
[tool.poetry]
name = "nonebot-plugin-styledstr"
version = "1.1.0-alpha.1"
description = "A styled string management plugin for NoneBot 2"
authors = ["Satoshi Jek <jks15satoshi@gmail.com>"]
license = "MIT"
readme = "README.md"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Typing :: Typed"
]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = ">=5.4.1,<7.0.0"
nonebot2 = "^2.0.0-alpha.16"
regex = { version = "^2021.11.10", optional = true }

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
yapf = "^0.31.0"
nonebug = "^0.1.0"

[tool.poetry.extras]
regex = ["regex"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
